Now allowing Assembly.idx to be a tuple
Assembly.idx is what identifies an assembly. Previously, it could only be a scalar. Now it can also be a tuple of any length.
For example, it is this way possible to identify an assembly by its seed and index within the seed.

Note: All assemblies within an AssemblyGroup must ALL have the same type of .idx, i.e. either all scalar or all tuple!

I also added a script to tests that checks that io is still working with this change.